korrigering i admin-katalog

+++ admin/public/main.js
@@ -41,18 +41,22 @@ let app = new Vue({
     },
     del: function (id) {
       // Ta bort recept
-      fetch(url + "/delete/" + id, {
-        //Lägger till IDt som skickades med som argument till url:en
-        method: "DELETE"
-      }).then(res => res.json()).then(data => console.log(data)).catch(err => console.log(err));
-      let found = -1; // Ta bort receptet ur recipeList-arrayen för att se aktuell receptlista utan att behöva ladda om sidan
+      if (!id) {
+        alert("Ladda om sidan innan borttagning kan genomföras.");
+      } else {
+        fetch(url + "/delete/" + id, {
+          //Lägger till IDt som skickades med som argument till url:en
+          method: "DELETE"
+        }).then(res => res.json()).then(data => console.log(data)).catch(err => console.log(err));
+        let found = -1; // Ta bort receptet ur recipeList-arrayen för att se aktuell receptlista utan att behöva ladda om sidan
 
-      for (let i = 0; i < this.recipeList.length; i++) {
-        if (id == this.recipeList[i]._id) {
-          found = i;
+        for (let i = 0; i < this.recipeList.length; i++) {
+          if (id == this.recipeList[i]._id) {
+            found = i;
 
-          if (found != -1) {
-            this.recipeList.splice(found, 1);
+            if (found != -1) {
+              this.recipeList.splice(found, 1);
+            }
           }
         }
       }
@@ -107,15 +111,19 @@ let app = new Vue({
       ob.image = updValueArr[5];
       let jsonStr = JSON.stringify(ob); // KOnvertera objektet till JSON
 
-      fetch(url + "/update/" + event.target.id, {
-        //Lägger till IDt till url:en
-        method: "PUT",
-        headers: {
-          "Content-Type": "application/json"
-        },
-        body: jsonStr // Skicka med JSON-strängen i anropet
+      if (!event.target.id) {
+        alert("Ladda om sidan innan uppdatering kan genomföras.");
+      } else {
+        fetch(url + "/update/" + event.target.id, {
+          //Lägger till IDt till url:en
+          method: "PUT",
+          headers: {
+            "Content-Type": "application/json"
+          },
+          body: jsonStr // Skicka med JSON-strängen i anropet
 
-      }).then(res => res.json()).then(data => console.log(data)).catch(err => console.log(err));
+        }).then(res => res.json()).then(data => console.log(data)).catch(err => console.log(err));
+      }
     },
     logout: function (event) {
       // Loggar ut admin när denne klickar på logga ut-knappen
